backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.9
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==24.3
django-timezone-field==7.2.1
djangorestframework==3.16.1
h11==0.16.0
idna==3.11
kombu==5.6.1
outcome==1.3.0.post0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-decouple==3.8.1
redis==5.0.8
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.4
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.9
django-celery-beat==2.8.1
django-celery-results==2.5.0
django-cors-headers==4.9.0
django-filter==24.3
django-timezone-field==7.2.1
djangorestframework==3.16.1
h11==0.16.0
idna==3.11
kombu==5.6.1
outcome==1.3.0.post0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
selenium==4.27.1
beautifulsoup4>=4.12.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.4
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
gunicorn
